Lint: Move CollectionExtensions (#12320)
* Lint: Move CollectionExtensions

* Forgot to change package name

* Missed a spot
