+++ core/src/com/unciv/logic/civilization/managers/QuestManager.kt
@@ -31,8 +31,8 @@ import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.models.translations.fillPlaceholders
 import com.unciv.models.translations.getPlaceholderParameters
 import com.unciv.models.translations.tr
-import com.unciv.ui.components.extensions.randomWeighted
 import com.unciv.ui.components.extensions.toPercent
+import com.unciv.utils.randomWeighted
 import kotlin.random.Random
 
 class QuestManager : IsPartOfGameInfoSerialization {
