+++ core/src/com/unciv/logic/civilization/managers/TechManager.kt
@@ -24,8 +24,8 @@ import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.MayaCalendar
-import com.unciv.ui.components.extensions.withItem
 import com.unciv.ui.components.fonts.Fonts
+import com.unciv.utils.withItem
 import kotlin.math.ceil
 import kotlin.math.max
 import kotlin.math.min
