+++ core/src/com/unciv/ui/popups/options/AdvancedTab.kt
@@ -35,6 +35,7 @@ import com.unciv.ui.screens.basescreen.BaseScreen
 import com.unciv.utils.Concurrency
 import com.unciv.utils.Display
 import com.unciv.utils.launchOnGLThread
+import com.unciv.utils.withoutItem
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
